Added an option to ignore all function names
GUI: Added an option to ignore all function names.
HEUR: The heuristic 'Same name' is ignore if the option "Ignore all function names" is selected.

In the future, it's possible that an embedded chooser will be shown to select which heuristics the analyst want to enable or disable. Not for now.