Bug fixes and changes to heuristics
GUI: Show specific messages when an option is automatically enabled or disabled.
GUI: Ratios are adjusted to show only 2 decimals.
HEUR: Both the timeout and the maximum number of rows processed when a cartesian product happens are adjusted according to the number of functions.
HEUR: When comparing pseudo-codes, clean them.
BUG: When comparing multiple lines of assembly, the assembly cleaning routine was wrong.
HEUR: Adjust the difference ratios to a maximum of 3 decimals.
GUI: Always show results with a ratio of 1.00 in the "Best Matches" list.
HEUR: Do a single pass for "Topological sort hash" instead of multiple ones.
HEUR: The minimum number of basic blocks to apply the heuristic "Strongly connected components" is set to 5
GUI: Results from many heuristics were not displayed in the correct list in some cases.