Initial version of 'Call Address Sequence' heuristic
HEUR: Added an initial version of the 'Call Address Sequence' heuristic. It's only available if experimental heuristics are enabled.
HEUR: Removed non-slow version of 'Strongly Connected Components' heuristic.
DB: Create an index for clean_asm and clean_pseudo too. In the future they will be removed as only a MD5 hash of them will be used.