Fixed a bug with underscored function names
BUG: Fixed a bug with function names with underscores and without them. If in a database there are both functions like "_bcopy" and "bcopy" the export process was incorrectly removing the first underscore, causing it a collision with another existing function.
GUI: Show better estimations of the time the export process takes.
EXPORT: Do not perform a commit when a hardcoded amount of functions were exported. Instead, perform a commit each time a 10% of the database's functions are exported (so, it will commit 10 times).
EXPORT: Also, do not refresh the waitbox every 1000 lines. Do it every 1% of the total database's functions (so, it will refresh 100 times).
Minor mods: Removed code with no purpose or redundant.