KOKA heuristic was wrong and other mods.
BUG: Results from the KOKA heuristic with a ratio smaller than 1.0 were put in the 'Best' results chooser.
GUI: As with Pigaios, if function names between binaries have true names (not auto-generated ones) and the names are different in both databases, show the result with a red background (because perhaps it's a false positive?).
GUI: If the call graphs are too different do not print hilarious messages like "Callgraphs from both programs differ in 300.343343434%".