Optimized various text search based heuristics causing problems with Windows and IDA 7.0/7.1
BUG: Optimized various queries that were causing hangs in Windows with IDA versions <= 7.1 when doing SQL queries searching for text without limiting the number of minimum nodes.
PARALLEL: Use at much cpu_count() - 1, as 1 thread is always being used (and seem to cause some hangs in Windows with IDA 7.0 and 7.1).
HEUR: Set back the timeout to 3 minutes instead of to 5 minutes as done recently. The timeout is being applied to each heuristic, not to the whole process.