Refactor do_export
Extract the list of functions exported by `do_export` into the
method `filter_functions`. This will allow a parallel version of
`do_export` to build a generator of functions to export on the fly
while receiving instruction from a main process.

Additionally, parallel export will not always export functions in
a predictable order. In the context of crashed exports, instead of
relying on the address of the last function inserted into the
database, all addresses added are retrieved and `self._funcs_cache`
is restored.